+++ test/fixture_factory.test.jl
@@ -32,32 +32,6 @@ end
 end
 
 
-# Return several values, create labels automatically
-@testcase "return values" begin
-    ff = Jute.fixture_factory(; returns_iterable=true, instant_teardown=true) do produce, v1, v2
-        produce([v1, v2])
-    end
-    val1 = MyType(1)
-    val2 = MyType(2)
-    lvals, rff = Jute.setup(ff, [val1, val2])
-    @test map(Jute.unwrap_value, lvals) == [val1, val2]
-    @test map(Jute.unwrap_label, lvals) == [string(val1), string(val2)]
-end
-
-
-# Return several values with custom labels
-@testcase "return values and labels" begin
-    ff = Jute.fixture_factory(; returns_iterable=true, instant_teardown=true) do produce, v1, v2
-        produce([v1, v2], ["one", "two"])
-    end
-    val1 = MyType(1)
-    val2 = MyType(2)
-    lvals, rff = Jute.setup(ff, [val1, val2])
-    @test map(Jute.unwrap_value, lvals) == [val1, val2]
-    @test map(Jute.unwrap_label, lvals) == ["one", "two"]
-end
-
-
 # Check that if delayed_teardown=true, teardown is not called right away
 @testcase "delayed teardown" begin
     teardown_started = false